requirements.txt
decorator==4.0.9
future==0.15.2
Mako==1.0.4
MarkupSafe==0.23
nose==1.3.7
python-telegram-bot==3.3
requests==2.9.1
six==1.10.0
validators==0.10
